pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "orchestrator-embrapii-srinfo"
version = "1.0.0"
description = "Orquestrador de Pipeline EMBRAPII SRInfo baseado em Prefect"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "Proprietary"}
authors = [
    {name = "EMBRAPII", email = "info@embrapii.org.br"}
]
dependencies = [
    "prefect>=2.13.0",
    "selenium",
    "webdriver-manager",
    "pandas",
    "openpyxl",
    "pywin32",
    "psutil",
    "python-dotenv",
    "clickhouse-connect",
    "gspread",
    "google-auth",
    "Office365-REST-Python-Client",
    "pymsteams",
]

[project.optional-dependencies]
dev = [
    "black",
    "isort",
    "flake8",
    "pytest",
    "pytest-cov",
]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
